- Added *AddressablesLoader* to the samples package with the asset loader implementation for the asset loader
**Changed**
- Removed the dependency to Unity's Addressables package
- Removed the Addressables helper methods from *AssetLoaderUtils* to remove any external package dependency
- Removed the Addressables editor build method
- Moved *AddressablesIdsGenerator* to the samples package
