--- Editor/GameLovers.AssetLoader.Editor.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 894425d9350e7744b9aba0e674e038bf
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
