+++ Runtime/GameLovers.AssetLoader.asmdef
@@ -1,10 +1,6 @@
 {
     "name": "GameLovers.AssetLoader",
-    "references": [
-        "GUID:9e24947de15b9834991c9d8411ea37cf",
-        "GUID:84651a3751eca9349aac36a66bba901b",
-        "GUID:22c6cdfa54ae844a9a9eda2f0014b020"
-    ],
+    "references": [],
     "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
