+++ Samples/AddressablesLoader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e787329d8e223ba4fb7666fc1b64d11b
+guid: c12cc1e510d684f278db67ae0b4f4b4c
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
