+++ Samples/AddressablesLoader/AddressablesLoader.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cb2a1051e4cfb6e4486dd92b365c7773
+guid: 0974504c9b7a14d79860d72f62469075
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
