+++ Samples/AddressablesLoader/Editor.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a5739548c2a94604bed8ffc51ac7cc4d
+timeCreated: 1595892245
