+++ Samples/AddressablesLoader/Editor/AddressableIdsGenerator.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c4ce5a9a733c485daae22a89dd2115d7
+timeCreated: 1595892255
