--- Samples/Example.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7269a1658a3cbcc4f98998dd155c00ab
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
