--- Samples/Example/.sample.json
@@ -1,5 +1,0 @@
-{
-	"displayName":"Example Sample",
-	"description": "Replace this string with your own description of the sample. Delete the Samples folder if not needed.",
-	"createSeparatePackage": false
-}
