--- Samples/Example/SampleExample.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 81dfc92af8b158440a9771b05767608a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
