+++ package.json
@@ -1,12 +1,16 @@
 {
   "name": "com.gamelovers.assetloader",
   "displayName": "Asset Loader",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "unity": "2019.3",
   "description": "This package extends the loading functionality of assets from Unity Addressables with the possibility to generate new info about each asset addressable.\n\nThe asset info generator can be executed in Tools > Generate AddressableIds",
-  "dependencies": {
-    "com.unity.addressables": "1.11.2"
-  },
   "type": "library",
-  "hideInEditor": false
+  "hideInEditor": false,
+  "samples": [
+    {
+      "displayName": "Addressables Loader",
+      "description": "Package implementation when using in combination with Addressables packages",
+      "path": "Samples~/AddressablesLoader"
+    }
+  ]
 }
